# docs/swagger.yaml
components:
  schemas:
...
    github_com_array_banking-api_internal_dto.ListTransactionsResponse:
      properties:
        pagination:
          $ref: '#/components/schemas/github_com_array_banking-api_internal_dto.PaginationInfo'
        transactions:
          items:
            $ref: '#/components/schemas/github_com_array_banking-api_internal_dto.TransactionWithBalance'
          type: array
          uniqueItems: false
      type: object
...
    github_com_array_banking-api_internal_dto.PaginationInfo:
      properties:
        hasMore:
          type: boolean
        limit:
          type: integer
        nextCursor:
          type: string
        total:
          type: integer
      type: object
...
    github_com_array_banking-api_internal_dto.TransactionRequest:
      properties:
        amount:
          type: string
        description:
          maxLength: 255
          minLength: 1
          type: string
        type:
          enum:
          - credit
          - debit
          type: string
      required:
      - amount
      - description
      - type
      type: object
    github_com_array_banking-api_internal_dto.TransactionWithBalance:
      properties:
        accountId:
          type: string
        amount:
          type: string
        category:
          type: string
        createdAt:
          type: string
        description:
          type: string
        id:
          type: string
        mccCode:
          type: string
        merchantName:
          type: string
        processedAt:
          type: string
        reference:
          type: string
        runningBalance:
          type: string
        status:
          type: string
        transactionType:
          type: string
      type: object
...
    github_com_array_banking-api_internal_errors.ErrorDetail:
      properties:
        code:
          type: string
        details:
          items:
            type: string
          type: array
          uniqueItems: false
        message:
          type: string
        trace_id:
          type: string
      type: object
    github_com_array_banking-api_internal_errors.ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorDetail'
      type: object
...
    github_com_array_banking-api_internal_models.JSONBMap:
      additionalProperties: {}
      description: Map of string keys to arbitrary values
      type: object
    github_com_array_banking-api_internal_models.Transaction:
      properties:
        account_id:
          type: string
        amount:
          type: number
        balance_after:
          type: number
        balance_before:
          type: number
        category:
          type: string
        created_at:
          type: string
        description:
          type: string
        id:
          type: string
        mcc_code:
          type: string
        merchant_name:
          type: string
        metadata:
          $ref: '#/components/schemas/github_com_array_banking-api_internal_models.JSONBMap'
        pending_until:
          type: string
        processed_at:
          type: string
        processing_fee:
          type: number
        reference:
          type: string
        reversal_reference:
          type: string
        reversed_at:
          type: string
        status:
          type: string
        transaction_type:
          type: string
        updated_at:
          type: string
        version:
          type: integer
      type: object
...
  securitySchemes:
    BearerAuth:
      description: Type "Bearer" followed by a space and JWT token.
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: Production-quality banking REST API for developer assessment and interviewing.
    Provides core banking functionality including identity management, account operations,
    customer management, and transaction processing.
  title: Array Banking API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /accounts/{accountId}/transactions:
    get:
      description: Retrieve paginated and filtered transaction history for a specific
        account with cursor-based pagination
      parameters:
      - description: Account ID (UUID)
        in: path
        name: accountId
        required: true
        schema:
          type: string
      - description: Pagination cursor for next page
        in: query
        name: cursor
        schema:
          type: string
      - description: Number of results per page (max 100)
        in: query
        name: limit
        schema:
          default: 20
          type: integer
      - description: Filter by start date (YYYY-MM-DD)
        in: query
        name: start_date
        schema:
          type: string
      - description: Filter by end date (YYYY-MM-DD)
        in: query
        name: end_date
        schema:
          type: string
      - description: Filter by transaction type
        in: query
        name: type
        schema:
          enum:
          - credit
          - debit
          type: string
      - description: Filter by status
        in: query
        name: status
        schema:
          enum:
          - pending
          - completed
          - failed
          - reversed
          type: string
      - description: Filter by category code
        in: query
        name: category
        schema:
          type: string
      - description: Filter by minimum amount
        in: query
        name: min_amount
        schema:
          type: string
      - description: Filter by maximum amount
        in: query
        name: max_amount
        schema:
          type: string
      - description: Filter by merchant name
        in: query
        name: merchant
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_dto.ListTransactionsResponse'
          description: Transaction history with pagination
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: VALIDATION_001 - Invalid parameters or VALIDATION_003 - Invalid
            account ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_005 - Account belongs to another user
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: ACCOUNT_001 - Account not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: List transactions
      tags:
      - Transactions
    post:
      description: Create a new transaction (credit or debit) on an account
      parameters:
      - description: Account ID (UUID)
        in: path
        name: accountId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/github_com_array_banking-api_internal_dto.TransactionRequest'
        description: Transaction details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_models.Transaction'
          description: Transaction created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: VALIDATION_001 - Invalid request body or account ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_005 - Account belongs to another user
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: ACCOUNT_001 - Account not found
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: TRANSACTION_002 - Invalid transaction amount, TRANSACTION_003
            - Insufficient funds, ACCOUNT_002 - Account not active
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Create a transaction
      tags:
      - Accounts
  /accounts/{accountId}/transactions/{id}:
    get:
      description: Retrieve detailed information about a specific transaction including
        running balance
      parameters:
      - description: Account ID (UUID)
        in: path
        name: accountId
        required: true
        schema:
          type: string
      - description: Transaction ID (UUID)
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_dto.TransactionWithBalance'
          description: Transaction details with running balance
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: VALIDATION_001 - Invalid transaction or account ID format
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_002 - Missing or invalid authentication
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: AUTH_005 - Transaction belongs to another user's account
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: ACCOUNT_001 - Account not found or TRANSACTION_001 - Transaction
            not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_array_banking-api_internal_errors.ErrorResponse'
          description: SYSTEM_001 - Internal server error
      security:
      - BearerAuth: []
      summary: Get transaction by ID
      tags:
      - Transactions
...